Fixes in the Ubuntu Installation script
The ub_version taken from lsb_release could be changed to simplify the
process. Changes to uv_v_major and *minor correctly split the ub_version
that comes as e. g. 22.04. Moved numpy installation to the major
controlled if statement to allow for the latest numpy library (allows
the usage of libraries such as pandas without the need for requirement
freezing)